Added API-gateway
Added Ocelot api gateway to the sollution. Removed unecicary endpoints. Started looking into integration tests

+++ RackApi.User/Controllers/UserController.cs
@@ -1,6 +1,7 @@
 ﻿using System.IdentityModel.Tokens.Jwt;
 using System.Security.Claims;
 using System.Text;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.IdentityModel.Tokens;
@@ -11,12 +12,11 @@ namespace RackApi.User.Controllers;
 
 [ApiController]
 [Route("[controller]")]
-
 public class UserController : ControllerBase
 {
     private readonly IConfiguration _configuration;
-    private readonly ILogger<UserController> _logger;
     private readonly ApiDbContext _context;
+    private readonly ILogger<UserController> _logger;
 
     public UserController(IConfiguration configuration, ILogger<UserController> logger, ApiDbContext context)
     {
@@ -25,23 +25,20 @@ public class UserController : ControllerBase
         _context = context;
     }
 
-    [HttpGet(Name = "Login")]
+    [HttpGet]
     public async Task<ActionResult<string>> Login(string email, string password)
     {
         var queryable = _context.Users.AsNoTracking();
 
         var reslut = await queryable.Where(x => x.Email == email && x.Password == password).FirstAsync();
 
-        if (reslut == null)
-        { 
-            return NotFound();
-        }
-        
-        var tokenString = generateJwtToken(reslut.Email);
+        if (reslut == null) return NotFound();
+
+        var tokenString = generateJwtToken(reslut.Id);
         return tokenString;
     }
 
-    [HttpPost(Name = "UpdateUser")]
+    [HttpPost]
     public async Task<ActionResult<UserModel>> Register(UserModel user)
     {
         user.CreatedAt = DateTime.Now;
@@ -49,44 +46,56 @@ public class UserController : ControllerBase
         _context.Users.Add(user);
         await _context.SaveChangesAsync();
 
-        var tokenString = generateJwtToken(user.Email);
-        return CreatedAtAction("Login", new { email = user.Email, password = user.Password }, user + " "+ tokenString);
+        var tokenString = generateJwtToken(user.Id);
+        return CreatedAtAction("Login", new { email = user.Email, password = user.Password }, tokenString);
     }
 
-    [HttpDelete(Name = "DeleteUser")]
+    [Authorize]
+    [HttpDelete]
     public async Task<ActionResult<string>> Delete(int id)
     {
+        var authorizationHeader = HttpContext.Request.Headers["Authorization"].FirstOrDefault();
+        string token = null;
+
+        if (!string.IsNullOrEmpty(authorizationHeader) && authorizationHeader.StartsWith("Bearer "))
+            token = authorizationHeader.Substring("Bearer ".Length).Trim();
+
+        var tokenHandler = new JwtSecurityTokenHandler();
+        var jwtToken = tokenHandler.ReadJwtToken(token);
+
+        var userId = jwtToken.Claims.FirstOrDefault(c => c.Type == "userId")?.Value;
+
+        if (id != Convert.ToInt16(userId)) return Unauthorized();
+
         var userToDelete = await _context.Users.FindAsync(id);
         if (userToDelete != null)
         {
             var producer = new RabbitMQProducer();
             producer.PublishMessage(id.ToString());
-            
+
             _context.Users.Attach(userToDelete);
             _context.Users.Remove(userToDelete);
             await _context.SaveChangesAsync();
             return Ok();
         }
-        else
-        {
-            return NotFound();
-        }
+
+        return NotFound();
     }
 
-    private string generateJwtToken(string email)
+    private string generateJwtToken(int id)
     {
         // Generate JWT Token
         var tokenHandler = new JwtSecurityTokenHandler();
         var key = Encoding.ASCII.GetBytes(_configuration["ConnectionStrings:DefaultJWTKey"]);
         var tokenDescriptor = new SecurityTokenDescriptor
         {
-            Subject = new ClaimsIdentity(new Claim[]
+            Subject = new ClaimsIdentity(new[]
             {
-                new Claim(ClaimTypes.Name, email),
-                // Add additional claims as needed
+                new Claim("userId", id.ToString())
             }),
-            Expires = DateTime.UtcNow.AddHours(1), // Token expiration time
-            SigningCredentials = new SigningCredentials(new SymmetricSecurityKey(key), SecurityAlgorithms.HmacSha256Signature),
+            Expires = DateTime.UtcNow.AddHours(8), // Token expiration time
+            SigningCredentials =
+                new SigningCredentials(new SymmetricSecurityKey(key), SecurityAlgorithms.HmacSha256Signature),
             Audience = "http://localhost:5114",
             Issuer = "http://localhost:5012"
         };
